// src/types.ts
export interface EnquiryForm {
  studentName: string;
  parentName: string;
  classLevel: string;
  subject: string;
  mobile: string;
  email: string;
  message: string;
}

export interface FeatureCardType {
  title: string;
  description: string;
  iconName: string;
}

export interface AboutPointType {
  title: string;
  description: string;
}

export interface ContactDetailType {
  title: string;
  value: string;
  iconName: string;
}